Simplify thank you message in Recipeasy.py
For any teacher who may be checking the commits. The reason it might appear so weird is because I coded everything and forgot to commit. And the commit was too large in the codespace on github to work. So I had to copy my code and go through and update everything manually.